Input console
Takes input and throws it up to an area to display it (click the enter
button instead of using enter key, when testing). If one of you’s can
test it on windows that would be great, because it looks nice for me,
but I’m not sure if it will for you’s.